[service] extract main file to a separate file
